+++ react/utils/definedMessages.ts
@@ -1,0 +1,21 @@
+import { defineMessages } from 'react-intl'
+
+export const input = defineMessages({
+  valuePlaceholder: { id: 'input.value.placeholder' },
+  valueLabel: { id: 'input.value.label' },
+  valueButton: { id: 'input.value.button' },
+  codePlaceholder: { id: 'input.code.placeholder' },
+  codeLabel: { id: 'input.code.label' },
+  codeButton: { id: 'input.code.button' },
+  tooltip: { id: 'input.tooltip' },
+})
+
+export const titles = defineMessages({
+  valueGiftCard: { id: 'titles.value.gift.card' },
+  money: { id: 'titles.money' },
+  redemption: { id: 'titles.redemption' },
+})
+
+export const wallet = defineMessages({
+  title: { id: 'wallet.title' },
+})