+++ CHANGELOG.md
@@ -5,4 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+- Initial store draft
+
+### Added
 - Initial store setup